requirements.txt
pywhatkit==5.4
Flask==2.3.3
PyGithub==1.59.0
apscheduler==3.10.1
requests==2.31.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
Werkzeug==2.3.7